refactor: move block/header builder for ephemeral bridge
