+++ bin/e2hs-writer/src/subcommands/mod.rs
@@ -1,5 +1,3 @@
-pub mod execution_block_builder;
-pub mod execution_header_builder;
 pub mod full_block;
 pub mod head_generator;
 pub mod single_generator;
