+++ bin/trin-execution/Cargo.toml
@@ -28,7 +28,7 @@ lazy_static.workspace = true
 parking_lot.workspace = true
 prometheus_exporter.workspace = true
 rand.workspace = true
-rayon = "1.10.0"
+rayon.workspace = true
 reqwest =  { workspace = true, features = ["stream"] }
 revm.workspace = true
 revm-inspectors = "0.20.0"
