+++ crates/ethportal-api/Cargo.toml
@@ -37,6 +37,7 @@ once_cell = "1.17"
 parking_lot.workspace = true
 quickcheck.workspace = true
 rand.workspace = true
+rayon.workspace = true
 rs_merkle = "1.4.2"
 secp256k1 = { version = "0.29.0", features = ["global-context", "recovery", "rand"] }
 serde = { workspace = true, features = ["rc"] }
