+++ crates/ethportal-api/src/types/execution/builders/mod.rs
@@ -1,0 +1,2 @@
+pub mod execution_block_builder;
+pub mod execution_header_builder;
