+++ crates/ethportal-api/src/types/execution/mod.rs
@@ -1,5 +1,6 @@
 pub mod accumulator;
 pub mod block_body;
+pub mod builders;
 pub mod ephermeral_header;
 pub mod header_with_proof;
 pub mod receipts;
